requirements.txt
langchain-openai
python-dotenv
langchain
langchain-community
langchain-anthropic
langchain_google_genai
langchain-google-firestore
firestore
chromadb
tiktoken
sentence-transformers
bs4
firecrawl-py
langchainhub
wikipedia
tavily-python
streamlit


